requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
pypdf
faiss-cpu
groq
langchain-groq
sentence_transformers